Quiz: refactor: move group notifications out of AssignmentManager
