+++ src/main/java/uk/ac/cam/cl/dtg/isaac/api/managers/AssignmentManager.java
@@ -27,49 +27,33 @@ import uk.ac.cam.cl.dtg.isaac.dao.IAssignmentPersistenceManager;
 import uk.ac.cam.cl.dtg.isaac.dto.AssignmentDTO;
 import uk.ac.cam.cl.dtg.isaac.dto.GameboardDTO;
 import uk.ac.cam.cl.dtg.segue.api.managers.GroupManager;
-import uk.ac.cam.cl.dtg.segue.api.managers.IGroupObserver;
-import uk.ac.cam.cl.dtg.segue.api.managers.UserAccountManager;
-import uk.ac.cam.cl.dtg.segue.api.managers.UserAssociationManager;
-import uk.ac.cam.cl.dtg.segue.auth.exceptions.NoUserException;
-import uk.ac.cam.cl.dtg.segue.comm.EmailManager;
-import uk.ac.cam.cl.dtg.segue.comm.EmailType;
 import uk.ac.cam.cl.dtg.segue.dao.SegueDatabaseException;
-import uk.ac.cam.cl.dtg.segue.dao.content.ContentManagerException;
 import uk.ac.cam.cl.dtg.segue.dos.GroupMembershipStatus;
 import uk.ac.cam.cl.dtg.segue.dto.UserGroupDTO;
 import uk.ac.cam.cl.dtg.segue.dto.users.GroupMembershipDTO;
 import uk.ac.cam.cl.dtg.segue.dto.users.RegisteredUserDTO;
-import uk.ac.cam.cl.dtg.segue.dto.users.UserSummaryWithEmailAddressDTO;
-import uk.ac.cam.cl.dtg.util.NameFormatter;
 import uk.ac.cam.cl.dtg.util.PropertiesLoader;
 
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
 import static uk.ac.cam.cl.dtg.segue.api.Constants.*;
-import static uk.ac.cam.cl.dtg.util.NameFormatter.getFilteredGroupNameFromGroup;
 import static uk.ac.cam.cl.dtg.util.NameFormatter.getTeacherNameFromUser;
 
 /**
  * AssignmentManager.
  */
-public class AssignmentManager implements IGroupObserver {
+public class AssignmentManager {
     private static final Logger log = LoggerFactory.getLogger(AssignmentManager.class);
 
     private final IAssignmentPersistenceManager assignmentPersistenceManager;
     private final GroupManager groupManager;
-    private final EmailManager emailManager;
     private final EmailService emailService;
-    private final UserAccountManager userManager;
-	private final GameManager gameManager;
+    private final GameManager gameManager;
     private final PropertiesLoader properties;
 
     /**
@@ -78,28 +62,20 @@ public class AssignmentManager implements IGroupObserver {
      *            - to save assignments
      * @param groupManager
      *            - to allow communication with the group manager.
-     * @param emailManager
-     *            - email manager
      * @param emailService
      *            - service for sending specific emails.
-     * @param userManager
-     *            - the user manager object
      * @param gameManager
  *            - the game manager object
      */
     @Inject
     public AssignmentManager(final IAssignmentPersistenceManager assignmentPersistenceManager,
-                             final GroupManager groupManager, final EmailManager emailManager,
-                             final EmailService emailService, final UserAccountManager userManager,
+                             final GroupManager groupManager, final EmailService emailService,
                              final GameManager gameManager, final PropertiesLoader properties) {
         this.assignmentPersistenceManager = assignmentPersistenceManager;
         this.groupManager = groupManager;
-        this.emailManager = emailManager;
         this.emailService = emailService;
-        this.userManager = userManager; 
-		this.gameManager = gameManager;
+        this.gameManager = gameManager;
         this.properties = properties;
-        groupManager.registerInterestInGroups(this);
     }
 
     /**
@@ -286,148 +262,6 @@ public class AssignmentManager implements IGroupObserver {
 
         return groups;
     }
-
-    @Override
-    public void onGroupMembershipRemoved(final UserGroupDTO group, final RegisteredUserDTO user) {
-		// do nothing
-    }
-
-    @Override
-    public void onMemberAddedToGroup(final UserGroupDTO group, final RegisteredUserDTO user) {
-		Validate.notNull(group);
-		Validate.notNull(user);
-
-        // Try to email user to let them know
-        try {
-            List<AssignmentDTO> existingAssignments = this.getAllAssignmentsForSpecificGroups(Collections.singletonList(group));
-
-            emailManager.sendTemplatedEmailToUser(user,
-                    emailManager.getEmailTemplateDTO("email-template-group-welcome"),
-                    this.prepareGroupWelcomeEmailTokenMap(user, group, existingAssignments),
-                    EmailType.SYSTEM);
-
-        } catch (ContentManagerException e) {
-            log.info("Could not send group welcome email ", e);
-        } catch (SegueDatabaseException e) {
-            log.error("Unable to send group welcome e-mail due to a database error. Failing silently.", e);
-        }
-    }
-
-    // TODO: This is not the right place for this.
-    @Override
-    public void onAdditionalManagerAddedToGroup(final UserGroupDTO group, final RegisteredUserDTO additionalManagerUser) {
-        Validate.notNull(group);
-        Validate.notNull(additionalManagerUser);
-
-        // Try to email user to let them know:
-        try {
-            RegisteredUserDTO groupOwner = this.userManager.getUserDTOById(group.getOwnerId());
-
-            String groupOwnerName = getTeacherNameFromUser(groupOwner);
-            String groupOwnerEmail = "Unknown";
-            if (groupOwner != null && groupOwner.getEmail() != null && !groupOwner.getEmail().isEmpty()) {
-                groupOwnerEmail = groupOwner.getEmail();
-            }
-            String groupName = "Unknown";
-            if (group.getGroupName() != null && !group.getGroupName().isEmpty()) {
-                groupName = group.getGroupName();
-            }
-
-            Map<String, Object> emailProperties = new ImmutableMap.Builder<String, Object>()
-                    .put("ownerName", groupOwnerName)
-                    .put("ownerEmail", groupOwnerEmail)
-                    .put("groupName", groupName)
-                    .build();
-            emailManager.sendTemplatedEmailToUser(additionalManagerUser,
-                    emailManager.getEmailTemplateDTO("email-template-group-additional-manager-welcome"),
-                    emailProperties,
-                    EmailType.SYSTEM);
-
-        } catch (ContentManagerException e) {
-            log.info("Could not send group additional manager email ", e);
-        } catch (NoUserException e) {
-            log.info(String.format("Could not find owner user object of group %s", group.getId()), e);
-        } catch (SegueDatabaseException e) {
-            log.error("Unable to send group additional manager e-mail due to a database error. Failing silently.", e);
-        }
-    }
-
-    /**
-     * Helper to build up the list of tokens for addToGroup email.
-     *
-     * @param userDTO - identity of user
-     * @param userGroup - group being added to.
-     * @param existingAssignments - Any existing assignments that have been set.
-     * @return a map of string to string, with some values that may want to be shown in the email.
-     * @throws SegueDatabaseException if we can't get the gameboard details.
-     */
-    private Map<String, Object> prepareGroupWelcomeEmailTokenMap(final RegisteredUserDTO userDTO, final UserGroupDTO userGroup,
-                                                                 final List<AssignmentDTO> existingAssignments)
-            throws SegueDatabaseException {
-        Validate.notNull(userDTO);
-
-        UserSummaryWithEmailAddressDTO groupOwner = userGroup.getOwnerSummary();
-        String groupOwnerName = getTeacherNameFromUser(groupOwner);
-
-        String teacherInfo;
-        if (!userGroup.getAdditionalManagers().isEmpty()) {
-            teacherInfo = String.format("your teachers %s and %s",
-                    userGroup.getAdditionalManagers().stream().map(NameFormatter::getTeacherNameFromUser).collect(Collectors.joining(", ")),
-                    groupOwnerName);
-        } else {
-            teacherInfo = String.format("your teacher %s", groupOwnerName);
-        }
-
-        String groupName = getFilteredGroupNameFromGroup(userGroup);
-
-        if (existingAssignments != null) {
-            existingAssignments.sort(Comparator.comparing(AssignmentDTO::getCreationDate));
-        }
-
-        final DateFormat DATE_FORMAT = new SimpleDateFormat("dd/MM/yyyy");
-        StringBuilder htmlSB = new StringBuilder();
-        StringBuilder plainTextSB = new StringBuilder();
-
-        if (existingAssignments != null && existingAssignments.size() > 0) {
-            htmlSB.append("Your teacher has assigned the following assignments:<br>");
-            plainTextSB.append("Your teacher has assigned the following assignments:\n");
-
-            for (int i = 0; i < existingAssignments.size(); i++) {
-                GameboardDTO gameboard = gameManager.getGameboard(existingAssignments.get(i).getGameboardId());
-                String gameboardName = existingAssignments.get(i).getGameboardId();
-                if (gameboard != null && gameboard.getTitle() != null && !gameboard.getTitle().isEmpty()) {
-                    gameboardName = gameboard.getTitle();
-                }
-
-                String gameboardUrl = String.format("https://%s/assignment/%s",
-                        properties.getProperty(HOST_NAME),
-                        existingAssignments.get(i).getGameboardId());
-
-                String dueDate = "";
-                if (existingAssignments.get(i).getDueDate() != null) {
-                    dueDate = String.format(", due on %s", DATE_FORMAT.format(existingAssignments.get(i).getDueDate()));
-                }
-
-                htmlSB.append(String.format("%d. <a href='%s'>%s</a> (set on %s%s)<br>", i + 1, gameboardUrl,
-                        gameboardName, DATE_FORMAT.format(existingAssignments.get(i).getCreationDate()), dueDate));
-
-                plainTextSB.append(String.format("%d. %s (set on %s%s)\n", i + 1, gameboardName,
-                        DATE_FORMAT.format(existingAssignments.get(i).getCreationDate()), dueDate));
-            }
-        } else if (existingAssignments != null) {
-            htmlSB.append("No assignments have been set yet.<br>");
-            plainTextSB.append("No assignments have been set yet.\n");
-        }
-
-        return new ImmutableMap.Builder<String, Object>()
-                .put("teacherName", groupOwnerName)
-                .put("teacherInfo", teacherInfo)
-                .put("groupName", groupName)
-                .put("assignmentsInfo", plainTextSB.toString())
-                .put("assignmentsInfo_HTML", htmlSB.toString())
-                .build();
-    }
-
     private List<AssignmentDTO> filterAssignmentsBasedOnGroupMembershipContext(List<AssignmentDTO> assignments, Long userId) throws SegueDatabaseException {
         Map<Long, Map<Long, GroupMembershipDTO>> groupIdToUserMembershipInfoMap = Maps.newHashMap();
         List<AssignmentDTO> results = Lists.newArrayList();
